energy-grid-env/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
numpy>=1.26.0
google-genai>=0.8.0
